pyproject.toml
[tool.poetry]
name = "test-django-backend"
version = "0.1.0"
description = ""
authors = ["RG1ee <niwows@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.2.2"
django-stubs = "^4.2.0"
djangorestframework-stubs = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.django-stubs]
django_settings_module = "src.upstudy.settings"
ignore_missing_model_attributes = true

[tool.poetry.scripts]
denleda = "src.manage:main"
